Rename util.lua to detour_util.lua
