+++ lua/detour_util.lua
